+++ hadoop-ozone/recon/src/test/java/org/apache/hadoop/ozone/recon/api/TestNSSummaryEndpointWithLegacy.java
@@ -370,7 +370,7 @@ public class TestNSSummaryEndpointWithLegacy {
     NSSummaryTaskWithLegacy nsSummaryTaskWithLegacy = 
         new NSSummaryTaskWithLegacy(reconNamespaceSummaryManager, 
                                     reconOMMetadataManager);
-    nsSummaryTaskWithLegacy.reprocess(reconOMMetadataManager);
+    nsSummaryTaskWithLegacy.reprocessWithLegacy(reconOMMetadataManager);
   }
 
   @Test
